requirements.txt
streamlit
torch
torchvision
scikit-learn
Pillow
numpy